+++ test/livecss.spec.js
@@ -1,0 +1,11 @@
+'use strict';
+
+var Live;
+
+Live = require("../lib/livecss");
+
+describe('livecss', function(){
+
+
+
+});